user_data/copilot_v1/requirements.txt
# Core dependencies
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.3.0
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-anthropic>=0.0.1
openai>=1.1.0
anthropic>=0.5.0
redis>=4.6.0
python-jose>=3.3.0  # For JWT
python-dotenv>=1.0.0
celery>=5.3.0
matplotlib>=3.7.0
pandas>=2.0.0
pandas-ta>=0.3.14b0
sqlalchemy>=2.0.0
aiohttp>=3.8.5
websockets>=11.0.0
python-multipart>=0.0.6  # For file uploads in FastAPI
llama-index>=0.8.0
